# Do not allocate extra bitvec for bitslice