--- a/src/bitmask.rs
+++ b/src/bitmask.rs
@@ -355,13 +355,23 @@ impl Bitmask {
     where
         F: FnOnce(u32) -> (PageId, BlockOffset),
     {
-        let mut bitvec = BitVec::<usize, Lsb0>::from_bitslice(bitslice);
+        // Get raw memory region
+        let (head, raw_region, tail) = bitslice
+            .domain()
+            .region()
+            .expect("Regions cover more than one usize");
+
+        // We expect the regions to not use partial usizes
+        debug_assert!(head.is_none());
+        debug_assert!(tail.is_none());
+
         let mut current_size: u32 = 0;
         let mut current_start: u32 = 0;
         let mut num_shifts = 0;
         // Iterate over the integers that compose the bitvec. So that we can perform bitwise operations.
         const BITS_IN_CHUNK: u32 = usize::BITS;
-        for (chunk_idx, &mut mut chunk) in bitvec.as_raw_mut_slice().iter_mut().enumerate() {
+        for (chunk_idx, chunk) in raw_region.iter().enumerate() {
+            let mut chunk = *chunk;
             while chunk != 0 {
                 let num_zeros = chunk.trailing_zeros();
                 current_size += num_zeros;
@@ -440,15 +450,14 @@ impl Bitmask {
             .domain()
             .region()
             .expect("Region covers more than one usize");
-        
+
         // We expect the region to not use partial usizes
         debug_assert!(head.is_none());
         debug_assert!(tail.is_none());
 
+        // Iterate over the integers that compose the bitslice. So that we can perform bitwise operations.
         let mut max = 0;
         let mut current = 0;
-
-        // Iterate over the integers that compose the bitslice. So that we can perform bitwise operations.
         const BITS_IN_CHUNK: u16 = usize::BITS as u16;
         let mut num_shifts = 0;
         for chunk in raw_region {
@@ -493,8 +502,17 @@ impl Bitmask {
             leading = max;
             trailing = max;
         } else {
-            leading = region.leading_zeros() as u16;
-            trailing = region.trailing_zeros() as u16;
+            leading = raw_region
+                .iter()
+                .take_while_inclusive(|chunk| chunk == &&0)
+                .map(|chunk| chunk.trailing_zeros())
+                .sum::<u32>() as u16;
+            trailing = raw_region
+                .iter()
+                .rev()
+                .take_while_inclusive(|chunk| chunk == &&0)
+                .map(|chunk| chunk.leading_zeros())
+                .sum::<u32>() as u16;
         }
 
         Gaps::new(leading, trailing, max)