Turns out, I hadn't read the `BitVec` documentation thoroughly and it IS indeed possible to read the bitslice directly.

Even now that we are copying each integer chunk when processing, not creating a new bitvec on the fly gives the following improvement:

```
Finished `bench` profile [optimized + debuginfo] target(s) in 2.17s
     Running benches/bitmask_bench.rs (target/release/deps/bitmask_bench-45d5d69951807178)
Gnuplot not found, using plotters backend
calculate_gaps          time:   [3.2224 µs 3.2324 µs 3.2428 µs]
                        change: [-10.470% -10.032% -9.6058%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

find_available_blocks   time:   [119.89 ns 120.66 ns 121.47 ns]
                        change: [-60.048% -59.627% -59.256%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe
```

### New numbers against RocksDB.

Before:
```
**read_heavy** with prefill_fraction 0.95
PayloadStorage:
1572864 operations across 1 thread(s) in 909.604709ms; time/op = 578ns
RocksDB:
1572864 operations across 1 thread(s) in 875.479625ms; time/op = 556ns

**insert_heavy** with prefill_fraction 0.0
PayloadStorage:
1572864 operations across 1 thread(s) in 2.756195458s; time/op = 1.752µs
RocksDB:
1572864 operations across 1 thread(s) in 3.359350125s; time/op = 2.135µs

**update_heavy** with prefill_fraction 0.5
PayloadStorage:
1572864 operations across 1 thread(s) in 2.001515917s; time/op = 1.272µs
RocksDB:
1572864 operations across 1 thread(s) in 1.950734792s; time/op = 1.24µs
```

After:
```
**read_heavy** with prefill_fraction 0.95
PayloadStorage:
1572864 operations across 1 thread(s) in 894.836625ms; time/op = 568ns
RocksDB:
1572864 operations across 1 thread(s) in 891.638333ms; time/op = 566ns

**insert_heavy** with prefill_fraction 0.0
PayloadStorage:
1572864 operations across 1 thread(s) in 1.89236525s; time/op = 1.203µs
RocksDB:
1572864 operations across 1 thread(s) in 3.440077292s; time/op = 2.187µs

**update_heavy** with prefill_fraction 0.5
PayloadStorage:
1572864 operations across 1 thread(s) in 1.601730208s; time/op = 1.018µs
RocksDB:
1572864 operations across 1 thread(s) in 2.04536125s; time/op = 1.3µs
```